Change colorscheme back to gruvbox ==================================
Changed away from gruvbox beacuse diagnostic messages was hard to read.
Tried tokyonight which was better but instead made line numbers hard to
read. After adjusting Windows terminal colors to match tokyonight I have
decided to give gruvbox another go, with terminal colors left on
tokyonight.

Have yet to disable the tokyonight plugin - although the colorscheme is
not set so it has no effect - so that it will be easy to swich between
them and evaluate.